♻️: 피드백 반영
