+++ src/components/About/DepartmentSection.jsx
@@ -1,0 +1,18 @@
+import styled from 'styled-components';
+import DepartmentIntro from './DepartmentIntro';
+
+const MainSection = styled.section`
+  margin: 7.5%;
+  display: flex;
+  flex-direction: column;
+`;
+
+function DepartmentSection() {
+  return (
+    <MainSection>
+      <DepartmentIntro />
+    </MainSection>
+  );
+}
+
+export default DepartmentSection;
