+++ src/components/Main/TitleSection.jsx
@@ -31,7 +31,8 @@ const IconFigcaption = styled.figcaption`
 function TitleSection() {
   return (
     <TitleContainer>
-      <TextGenerator text="Let's Lead The Industry To A Broader FIELD" font='Goblin One' />
+      <TextGenerator text="Let's Lead The Industry" font='Goblin One' />
+      <TextGenerator text='To A Broader FIELD' font='Goblin One' />
       <Figure $position='absolute'>
         <Icon src='scrollDown.png' />
         <IconFigcaption>아래로 스크롤하세요</IconFigcaption>
